requirements.txt
discord.py
arrow

